♻️ clean code

+++ packages/backend/user/src/user.controller.ts
@@ -6,8 +6,10 @@ import { IUserCreateResponse } from './interfaces/user-create-response.interface
 import { IUserDeleteResponse } from './interfaces/user-delete-response.interface'
 import { IUserForgotPasswordResponse } from './interfaces/user-forgot-password-response.interface'
 import { IUserGetByLinkResponse } from './interfaces/user-get-by-link-response.interface'
+import { IUserLink } from './interfaces/user-link.interface'
 import { IUserListParams } from './interfaces/user-list-params.interface'
 import { IUserListResponse } from './interfaces/user-list-response.interface'
+import { IUserResendResponse } from './interfaces/user-resend-response.interface'
 import { IUserSearchResponse } from './interfaces/user-search-response.interface'
 import { IUserUpdateByIdResponse } from './interfaces/user-update-by-id-response.interface'
 import { IUserUpdateParams } from './interfaces/user-update-params.interface'
@@ -29,9 +31,7 @@ export class UserController {
     let result: IUserSearchResponse
 
     if (searchParams.email && searchParams.password) {
-      const user = await this.userService.searchUser({
-        email: searchParams.email
-      })
+      const user = await this.userService.searchUserByEmail(searchParams.email)
 
       if (user) {
         if (await user.compareEncryptedPassword(searchParams.password)) {
@@ -153,13 +153,12 @@ export class UserController {
     link: string
   }): Promise<IUserConfirmResponse> {
     let result: IUserConfirmResponse
-
     if (confirmParams) {
       const userLink = await this.userService.getUserLink(confirmParams.link)
 
       if (userLink) {
-        const userId = userLink.user_id
-        const user = await this.userService.updateUserById(userId, {
+        const user = await this.userService.updateUserById(userLink.user, {
+          last_login: new Date(),
           password: confirmParams.password,
           is_confirmed: true
         })
@@ -197,9 +196,9 @@ export class UserController {
     email: string
   }): Promise<IUserForgotPasswordResponse> {
     let result: IUserForgotPasswordResponse
-    const user = await this.userService.searchUser({
-      email: forgotPasswordParams.email
-    })
+    const user = await this.userService.searchUserByEmail(
+      forgotPasswordParams.email
+    )
     if (user) {
       const userLink = await this.userService.createUserLink(
         user.id,
@@ -218,7 +217,8 @@ export class UserController {
           context: {
             name: user.name,
             email: user.email,
-            link: this.userService.getConfirmationLink(userLink.link)
+            link: this.userService.getConfirmationLink(userLink.link),
+            site: this.userService.getWebUrl()
           }
         })
         .toPromise()
@@ -227,6 +227,64 @@ export class UserController {
         status: HttpStatus.NOT_FOUND,
         message: 'user_forgot_password_not_found',
         errors: null
+      }
+    }
+
+    return result
+  }
+
+  @MessagePattern('user_resend')
+  public async resendMail(id: string): Promise<IUserResendResponse> {
+    let result: IUserResendResponse
+    let userLink: IUserLink
+    const user = await this.userService.searchUserById(id)
+
+    if (user) {
+      if (user.is_confirmed) {
+        userLink = await this.userService.createUserLink(
+          user.id,
+          Date.now() + 2 * 60 * 60 * 1000
+        )
+        this.mailerServiceClient
+          .send('mail_send', {
+            to: user.email,
+            subject: 'Recuperação de senha',
+            template: '/templates/forgot_password',
+            context: {
+              name: user.name,
+              email: user.email,
+              link: this.userService.getConfirmationLink(userLink.link),
+              site: this.userService.getWebUrl()
+            }
+          })
+          .toPromise()
+      } else {
+        userLink = await this.userService.getUserLinkByUser(user.id)
+        if (!userLink) {
+          userLink = await this.userService.createUserLink(user.id)
+        }
+        this.mailerServiceClient
+          .send('mail_send', {
+            to: user.email,
+            subject: 'E-mail de Confirmação',
+            template: '/templates/confirm_email',
+            context: {
+              name: user.name,
+              email: user.email,
+              link: this.userService.getConfirmationLink(userLink.link),
+              site: this.userService.getWebUrl()
+            }
+          })
+          .toPromise()
+      }
+      result = {
+        status: HttpStatus.OK,
+        message: 'user_resend_success'
+      }
+    } else {
+      result = {
+        status: HttpStatus.NOT_FOUND,
+        message: 'user_resend_not_found'
       }
     }
 
@@ -244,9 +302,9 @@ export class UserController {
           userParams.personal_data.cpf &&
           userParams.personal_data.dob))
     ) {
-      const usersWithEmail = await this.userService.searchUser({
-        email: userParams.email
-      })
+      const usersWithEmail = await this.userService.searchUserByEmail(
+        userParams.email
+      )
 
       if (usersWithEmail) {
         result = {
@@ -255,20 +313,36 @@ export class UserController {
           data: null,
           errors: {
             email: {
-              message: 'Email already exists',
+              message: 'E-mail already registered',
               path: 'email'
             }
           }
         }
       } else {
         try {
-          userParams.is_confirmed = false
-          userParams.password = Math.random()
-            .toString(36)
-            .replace(/[^a-z]+/g, '')
+          if (userParams.role !== 'PARTICIPANT') {
+            userParams.is_confirmed = false
+            userParams.password = Math.random().toString(16).replace('0.', '')
+          } else {
+            const usersWithCPF = await this.userService.searchUserByCPF(
+              userParams.personal_data.cpf
+            )
+            if (usersWithCPF) {
+              return {
+                status: HttpStatus.CONFLICT,
+                message: 'user_create_conflict',
+                data: null,
+                errors: {
+                  cpf: {
+                    message: 'CPF already registered',
+                    path: 'cpf'
+                  }
+                }
+              }
+            }
+          }
 
           const createdUser = await this.userService.createUser(userParams)
-          const userLink = await this.userService.createUserLink(createdUser.id)
           delete createdUser.password
           result = {
             status: HttpStatus.CREATED,
@@ -276,18 +350,24 @@ export class UserController {
             data: { user: createdUser },
             errors: null
           }
-          this.mailerServiceClient
-            .send('mail_send', {
-              to: createdUser.email,
-              subject: 'Email de Confirmação',
-              template: '/templates/confirm_email',
-              context: {
-                name: createdUser.name,
-                email: createdUser.email,
-                link: this.userService.getConfirmationLink(userLink.link)
-              }
-            })
-            .toPromise()
+          if (userParams.role !== 'PARTICIPANT') {
+            const userLink = await this.userService.createUserLink(
+              createdUser.id
+            )
+            this.mailerServiceClient
+              .send('mail_send', {
+                to: createdUser.email,
+                subject: 'E-mail de Confirmação',
+                template: '/templates/confirm_email',
+                context: {
+                  name: createdUser.name,
+                  email: createdUser.email,
+                  link: this.userService.getConfirmationLink(userLink.link),
+                  site: this.userService.getWebUrl()
+                }
+              })
+              .toPromise()
+          }
         } catch (e) {
           result = {
             status: HttpStatus.PRECONDITION_FAILED,
@@ -319,8 +399,48 @@ export class UserController {
       try {
         const user = await this.userService.searchUserById(params.id)
         if (user) {
+          if (user.personal_data.cpf)
+            params.user.personal_data.cpf = user.personal_data.cpf
+
           const updatedUser = Object.assign(user, params.user)
+
+          if (params.user.email && params.user.email !== user.email) {
+            const usersWithEmail = await this.userService.searchUserByEmail(
+              params.user.email
+            )
+            if (params.user.email && usersWithEmail) {
+              return {
+                status: HttpStatus.CONFLICT,
+                message: 'user_update_by_id_conflict',
+                user: null,
+                errors: {
+                  email: {
+                    message: 'E-mail already registered',
+                    path: 'email'
+                  }
+                }
+              }
+            }
+            updatedUser.is_confirmed = false
+            updatedUser.password = Math.random().toString(16).replace('0.', '')
+          }
           await updatedUser.save()
+          if (params.user.email) {
+            const userLink = await this.userService.createUserLink(user.id)
+            this.mailerServiceClient
+              .send('mail_send', {
+                to: user.email,
+                subject: 'E-mail de Confirmação',
+                template: '/templates/confirm_email',
+                context: {
+                  name: user.name,
+                  email: user.email,
+                  link: this.userService.getConfirmationLink(userLink.link),
+                  site: this.userService.getWebUrl()
+                }
+              })
+              .toPromise()
+          }
           result = {
             status: HttpStatus.OK,
             message: 'user_update_by_id_success',
@@ -336,6 +456,8 @@ export class UserController {
           }
         }
       } catch (e) {
+        console.log(e)
+
         result = {
           status: HttpStatus.PRECONDITION_FAILED,
           message: 'user_update_by_id_precondition_failed',
@@ -358,12 +480,16 @@ export class UserController {
   @MessagePattern('user_delete_by_id')
   public async userDeleteForUser(params: {
     id: string
+    participant?: boolean
   }): Promise<IUserDeleteResponse> {
     let result: IUserDeleteResponse
 
     if (params && params.id) {
       try {
         const user = await this.userService.searchUserById(params.id)
+
+        if (params.participant && user.role !== 'PARTICIPANT')
+          throw new Error('Not allowed')
 
         if (user) {
           await this.userService.removeUserById(params.id)
